SCRD-1964 Vertica 9 install fails on ironic devenv
The Vertica 9 installer checks for cloud-init data being available, via
169.254.169.254, and if detects it, it makes the assumption that you are
running in an AWS instance, but then fails to detect the AWS instance
type since the cloud-init data is an OpenStack flavour.

When trying to deploy a Vagrant cloud on an Ironic DevEnv setup this is
undersirable behaviour since it is detecting the cloud-init data for the
Ironic node itself, thanks to the request from the Vagrant VM being
NAT'd to the host environment.

To get around this we need to specify two additional options to the
install_vertica script:
  * --ignore-install-config
    This hidden option is used to disable the use of cloud-init data
    for configuring the install process

  * --ignore-aws-instance-type
    This option is needed to disable the check for AWS instance type
    which seems to invoked, and fail, even if --ignore-install-config
    is specified.

Ported-From: Iffd33aae2c476f7b7b99754fb984a2986fed4ca2
(cherry picked from https://gerrit.suse.provo.cloud/1880)

Change-Id: I36cc3558a6ced37e4a042036f78809b005869dcf