SCRD-2385 Use OS-dependent Apache2 vhost directory
On SUSE operating systems it is more natural to place all enabled
virtual host config files in `/etc/apache2/vhosts.d`, as opposed to
Debian-style configuration which places virtual host configs in a
staging area and then enables them with a utility called `a2ensite`
which creates a symlink between the staging area and the active
config directory. Up until now we've been forcing the Debian-style
method SLES since historically that is what Ardana has been built for.
This patch uses the new `apache2_vhost_dir` fact set by the `FND-AP2`
role (in cluster-ansible) to put vhost configuration in the right place
depending on the host OS, and stops using the `a2ensite` command for
SUSE hosts.

This patch also removes a redundant task that places the monasca-api
vhost in the directory immediately after another task that does the
same.

Depends-On: I9c739dceb87055263b1d7903ebb5b8d617b4dcd3
Change-Id: Ibfc74d7404f9c6d4e6d05857b03f42885372cf37