SCRD-1458 Fix monasca-api initial log permissions
When monasca-api is configured with the java implementation, the
monasca-api.log log is `touch`'d so that it is initialized with the
correct permissions. With the python implementation, this does not
happen, which is a problem because the service will create it with the
wrong group and cause problems for logrotate. This patch copies the log
file setup from the configure.yml task to the configure-pythonimpl.yml
task so that the two implementations are the same.

This also removes an unnecessary task which initialized the
apache-related monasca-api log files to the wrong group. The task
immediately following it was setting the right permissions.

Change-Id: Id3bb963850e09608ee904b9e6639ae771a7db58c