Re-introduce monasca-agent deps installation (bsc#1082403)
The install task is needed for both venv and package based installation
on SLES and RHEL, as it makes e.g. sure that setcap (required by the
libvirt plugin) is present on the system.

This also reduced some code duplication by using the same code path for
RHEL and legacy builds (both install the venv).

Change-Id: I34d93e05faed64b2db1bc006763fe5e0b88e06d0