SCRD-2652 Changes for supporting monasca-agent venv on RHEL compute
To install monasca-agent venv on RHEL, service file needs to be copied
to expected systemd location. Also creating expected path for storing
process pid and handling reboot cases.

Modifying monasca-agent-plugin library code to lookup for monasca-setup
in /usr/local/bin/ path as well. This is needed in case of RHEL where
venv is used but bin files are now stored in /usr/local/bin instead of
/opt/monasca/bin as done in legacy builds.

Creating symlink for monasca-setup to venv usr/bin/ path.

Adding symlink for venv/usr/share to venv/share for monasca agent template
lookup to work. This is related to the way its venv is created which is
different from legacy venv creation.

Change-Id: I90edfd0fc454d1bcd78c2753605ec8cf920ab589