Use --delete-config to avoid errors in journal (bsc#1087666)
Our version of kafka claims to be 0.9.0.1, but the behavior
doesn't match the Apache documentation.
http://kafka.apache.org/090/documentation.html
https://kafka.apache.org/0101/documentation.html

The 0.9.0 documentation specifies --deleteConfig, but based
on behavior this needs to be changed to --delete-config for
what we are deploying. Also, kafka-topics.sh has been
deprecated for changing topic configuration (including delete)
so switch to kafka-configs.sh.

Change-Id: I5a2f4fd2c93895127d8155424b7401a66eb18eec