Don't change the monasca- users if in use (bsc#1092182)
It was found in update testing that the ansible script was attempting
to change the monasca-[persister, thresh, notification] user but there
are processes running as the monasca- user, which caused a failure.

This change catches the error from the attempt to change the existing
user and if the failure is due to existing processes under the
monasca- user it continues on.

The advantage of this approach is that if the monasca- processes are
stopped before the install.yml is run then the change can be made.

Change-Id: If8c02577a60cdab5250ae8f778141e564f89b56b