Remove recursively owning /etc/monasca by mon-api (bsc#1094873)
The configure-pythonimpl.yml is creating the 3 configuration
files needed for monasca-api, then had a task to set ownership
recursively on the /etc/monasca directory.  This causes some
slopiness, especially if other services have installed first.
In the bug report, monasca-agent was installed first and had
its configurations changed to the mon-api user.

The simple approach is to remove that task, which doesn't seem
needed when the 3 config files are already created as owned by
mon-api.

Additionally, for non-legacy installs the configure.yml for
monasca-agent should have permissions for /etc/monasca/agent
that match what is set by the rpm (monasca-agent:monasca, 750).

Change-Id: If6b9e62e0307a2e6264f15b94b9ef7e451ec50d5