SCRD-5919 Optimise monasca-agent role
Issue: SCRD-5919

Redirect loading of tasks from correct playbooks based on run_mode var.

Other roles depend on monasca-agent role for monasca_agent_plugin.
This dependence causes monasca-agent role to be evaluated everytime the
these other-roles are loaded.

We skip executing any of the tasks from the monasca-agent role by using
'run_mode: Use'. This is good but as a side effect ansible still goes
ahead and evaluates every taskfile referred by main.yml in
monasca-agnet/tasks. Thus there ever is a performance hit. This hit is
not major if done only once. But since almost every role depends on
monasca-agent the cost of uselessly evaluating these playbooks again and
again is quite high and also ends up polluting the log file.

One alternative solution is to move the monasca_agent_plugin to the base
dir and remove this dependence entirely, but the problem with this
alternate approach is, that some of the other-roles use variables exported by
monasca-agent role. We could redefine those vars in these other-roles.
But that is potential bug: two-sources-of-truth

So this approach where the actual set of tasks depend on the variable
run_mode which in case of run_mode:Use is a noop.

Change-Id: I1c33a8ad1fdf79b2f0a64c4596a15e354a67f3a1