Don't modify /usr/lib/monasca/agent/ custom dirs (bsc#1094873)
The rpm installs the /usr/lib/monasca/agent/custom_checks.d and
custom_detect.d directories, so let it handle permissions.

Additionally, the change from commit
04b66d4f770713e33e025fff7b74bb2998068ce1 was overcorrecting
and it is not needed to change those dir to monasca ownership.
To be consistent with other /usr/lib/ dirs, the permissions should
be 755 root:root which the .rpm sets (once change
https://review.openstack.org/#/c/604847/ lands upstream).

Change-Id: I51ee22b471c114168ae5c7157ae5ecdf2afef34c