SCRD-6658 Add missing run_mode task files
Isses: SCRD-6658

The recent optimisation work failed to create run_mode files for the
Stop and Logging modes.

For Stop, list start, we just want to run the stop.yml tasks. For
Logging, the original code didn't do anything for that mode, so
similar to the Use mode, the Logging mode file is a no-op debug
tasks. Updated the run_mode_use.yml to reflect that the Use mode
is actually used in many places and not just to satisfy the code
linter.

Change-Id: I75cccf91c5c654f87abec454f25bc1ef6eafc7bc