SCRD-8482 Use correct path for monasca_db command (SCRD-8482)
The monasca_api_bin_dir ansible fact should specify the correct path
under which we can find the monasca_db command, so use that when we
run the upgrade action.

The incorrect use of the /opt/stack/service/monasca-api-*/ path results
in a shell expansion that includes multiple items when there are more
that one venvs incarnations installed as is the case for when we are
testing the update workflow.

Change-Id: I6338d7f331fee8d5b70f8aa1f629ea74a8f995c8