Stop existing mon-api services during upgrade (SOC-10470)
Stop services running as mon-api before attempting to update the mon-api
user settings.

With this patch the existing services will be stopped if called for a
stop or upgrade run_mode.

Note that technically we only appear to need this change in the Python
implementation stream but it is included in the Java stream as well for
completeness reasons.

Change-Id: Ie63a8b949202b04d69f510aab44b2d0c94940db8