align Monasca DB schema with upstream prior to upgrade (SOC-10277)
Prior to stable/rocky, Monasca DB schema was a wild, wild, west.
In other words, there was no "official version" of the schema. It was
left up to the vendors. And guest what, we have our own version of the
stable/pike schema! In stable/rocky, upstream Monasca established an
official version of the schema, and it was stamp (or fingerprinted) with
each migration step. 'f69cb3152a76' was the last known fingerprint in
stable/rocky so we'll need fix up our schema to align with the official
stable/rocky version prior to the migration. Otherwise, running
'monasca_db upgrade' will blow up on our face like a stink bomb.

Change-Id: Ic996181bbcc715d8e22706c007fe87ab261849d3