Leverage schema conversion script for upgrade (SOC-10277)
Create a new monasca-common role in the ansible code base to host the
mysql_schema_dir and monasca_mysql_schema_sync_script_file settings and
ensure that that role is included in all other relevant roles, such as
the monasca-api role, which is where the new conversion script needs to
be run from.

Similarly move the invocation of the new schema migration process into
the monasca-api role, where we trigger it just prior to the monasca_db
upgrade command to ensure that the Monasca DB has been migrated and
stamped apropriately.

Switch to using command rather than shell action when invoking command
lines that don't require require any shell features.

Change-Id: I7b8234c44a80a308783ff067a5cb6f29049818ee