Retry systemctl status for auto-restarting services (SOC-11210)
This patch implements some retry handling that will retry running
systemctl status for monasca-ansible managed services when their
current reported "Active:" line's status is not one of:
  * active (exited)
  * active (running)
  * inactive (dead)

We use the new Jinja2 filter plugin, systemctl_active_or_dead, to
check the registered result of the systemctl status command; this
was needed as attempting to use a complex until condition in the
ansible playbook was not viable.

If one of the above statuses is detected, the retry loop will be
skipped and normal execution of the ansible tasks, which will fail
out if the result of the systemctl status indicates that the service
is not healthy. Otherwise we will wait until the retries have been
exhausted and again take appropriate action based upon the final
systemctl status for the service.

The net effect of this approach is that we will proceed promptly
when the service status is healthy, or fail quickly if the service
status is failed/dead, but will wait for a certain amount of time
to allow auto-restarting services to start up successfully if they
encounter an ephemeral issue that requires a restart.

The retry count, and number of second delay between each retry are
specified by these variables:
  * monasca_service_status_retry_count, default 60
  * monasca_service_status_retry_delay, default 2

These values can be overriden when including the service_status.yml
task if values other than the defaults are desired.

Change-Id: I9b8e9f48299f86c60139f32062536192dba960b0