monasca-thresh restart and storm upgrade enhancements (SOC-11209)
Update the monasca-wrapper-thresh script to better handle storm list
hangs by using timeout to kill off the running command if it doesn't
complete within the specified timeout, which defaults to 60 seconds.

Update the monasca-thresh service startup handling to leverage the
recent changes to the monasca service status checking to determine
if the monasca-thresh service has started successfully or not.

Also simplify and fix the guard conditions for the determination of
whether to stop and restart various storm related services during
an upgrade.

Change-Id: I48978f23585a02994c0eecaca0e9773ef1ded73d