Fix monasca-thresh-wrapper status action (SOC-11209)
Fix storm_svc_check to return with the status from a failing timeout
or the grep for thresh-cluster in the output from storm list.

The previous re-working of the monasca-thresh-wrapper script added
support for timing out the storm command when it was being invoked
with the list action, to avoid problems with it hanging forever,
with a default timeout of 60 seconds.

However it appears that in some cases the 60s timeout may not be
enough so this patch increases it to 2 minutes.

Also cleanup the timeout override handling in the storm_check_svc()
shell function in monasca-thresh-wrapper.

Change-Id: I1040f403fa06a5e7f24fb08232688a992d2abd06