Fix stop start/stop logic (SOC-11209)
We are failing to stop minus on all nodes where it in running during
an upgrade from ardana soc8 to soc9, this was causing
problem with bring the stopped node back into the storm cluster

1: monasca-status.yml We were only stopping nimbus on the first member
but not any other members. This was hold over from the initial
iplementaion where nimbus only ran on the first node.

2:monasca-stop.yml  remove check for old Storm since we will only be upgrading
on systems that are post 0.9.5

3: roles/storm/tasks/main.yml cleanup install when checks to be more readable

4: stop_on_upgrade
   change thresh stop check to be on thresh not nimbus
   remove unneeded when chacks

5: .gitreview
change host to new FQDM .ci.prv.suse.net
We don't want to do an on mass change for this since it would cause churn in the
MU's but is good to do in conjuction with other changes.

Change-Id: I0ce82d06e35528f3eeaae51c1604d82217aba118